requirements.txt
# configuration
hydra-core

# pytorch
pytorch-ignite
torch
torchvision
torchaudio

# experiment tracking
tqdm
wandb